WIP: Full layout rework
