+++ launcher/ui/__init__.py
@@ -1,5 +1,13 @@
-__all__ = ["SettingsDialog", "Login", "NewsArea", "NewsTab", "ServerStatusBar"]
+__all__ = [
+    "SettingsDialog",
+    "Login",
+    "NewsArea",
+    "NewsTab",
+    "ServerStatusBar",
+    "Background",
+]
 
+from launcher.ui.background import Background
 from launcher.ui.login import Login
 from launcher.ui.news_area import NewsArea, NewsTab
 from launcher.ui.server_status_bar import ServerStatusBar
