+++ launcher/ui/server_status_bar.py
@@ -1,22 +1,22 @@
 from PySide6 import QtCore
-from PySide6.QtCore import QTimer
-from PySide6.QtWidgets import (
-    QGridLayout,
-    QHBoxLayout,
-    QLabel,
-    QSizePolicy,
-    QSpacerItem,
-    QWidget,
-)
+from PySide6.QtCore import Qt, QTimer
+from PySide6.QtGui import QColor, QPainter, QPen
+from PySide6.QtWidgets import QGridLayout, QLabel, QSizePolicy, QWidget
 
 from launcher import threads
+from launcher.ui import helpers
 
 
 class ServerStatusBar(QWidget):
     def __init__(self, parent=None):
         super().__init__(parent)
-        layout = QHBoxLayout()
         server_layout = QGridLayout()
+
+        self.setAutoFillBackground(True)
+        p = self.palette()
+        p.setColor(self.backgroundRole(), QColor(27, 47, 78, 220))
+        self.setPalette(p)
+
         server_label_style = """
             QLabel {
                 color: white;
@@ -28,7 +28,7 @@ class ServerStatusBar(QWidget):
 
         status_label_style = """
             QLabel {
-                color: green;
+                color: yellow;
                 font-size: 16px;
                 margin-bottom: 2px;
                 font-weight: bold;
@@ -41,19 +41,90 @@ class ServerStatusBar(QWidget):
         login_server_label = QLabel("Login Server:")
         login_server_label.setStyleSheet(server_label_style)
 
-        self.game_server_status_label = QLabel()
+        self.game_server_status_label = QLabel("CHECKING")
         self.game_server_status_label.setStyleSheet(status_label_style)
 
-        self.login_server_status_label = QLabel()
+        self.login_server_status_label = QLabel("CHECKING")
         self.login_server_status_label.setStyleSheet(status_label_style)
+
+        daily_reset_label = QLabel()
+        daily_reset_label.setStyleSheet(server_label_style)
+
+        daily_reset_timer_label = QLabel()
+        daily_reset_timer_label.setStyleSheet(server_label_style)
+
+        weekly_reset_label = QLabel()
+        weekly_reset_label.setStyleSheet(server_label_style)
+
+        weekly_reset_timer_label = QLabel()
+        weekly_reset_timer_label.setStyleSheet(server_label_style)
+
+        import datetime
+
+        import pytz
+
+        # Set the target timezone to UTC-7
+        target_tz = pytz.timezone("America/Los_Angeles")
+
+        local_tz = datetime.datetime.now(pytz.timezone("UTC")).astimezone().tzinfo
+
+        # Get the current local time
+        now_local = datetime.datetime.now(local_tz)
+
+        # Convert to UTC
+        now_utc = now_local.astimezone(pytz.utc)
+
+        # Set the time to 5:00 am in UTC-7
+        target_time = datetime.time(5, 0)
+        target_dt = datetime.datetime.combine(now_utc.date(), target_time)
+        target_dt += datetime.timedelta(days=1)
+        # Convert to the target timezone
+        target_dt_tz = target_tz.localize(target_dt)
+
+        # Convert back to the local timezone
+        target_dt_local = target_dt_tz.astimezone(local_tz)
+
+        next_weekday = target_dt_local.weekday()
+        next_weekday_name = datetime.date(1900, 1, next_weekday + 1).strftime("%a")
+
+        # Calculate the time difference
+        time_diff = target_dt_local - now_local
+
+        # Get the total number of minutes until the target datetime
+        total_minutes = int(time_diff.total_seconds() / 60)
+
+        # Calculate the number of hours and minutes
+        hours = total_minutes // 60
+        minutes = total_minutes % 60
+
+        # Combine the weekday name and time
+        # result = f"{next_weekday_name} {time_str}"
+        daily_reset_label.setText(
+            f"Daily Reset: {next_weekday_name} {target_dt_local.strftime('%H:%M')}"
+        )
+        daily_reset_timer_label.setText(f"{hours:02d}h {minutes:02d}m until reset")
+
+        weekly_reset_label.setText(
+            f"Weekly Reset: {next_weekday_name} {target_dt_local.strftime('%H:%M')}"
+        )
+        weekly_reset_timer_label.setText(f"{hours:02d}h {minutes:02d}m until reset")
 
         server_layout.addWidget(game_server_label, 0, 0)
         server_layout.addWidget(self.game_server_status_label, 0, 1)
         server_layout.addWidget(login_server_label, 1, 0)
         server_layout.addWidget(self.login_server_status_label, 1, 1)
+        server_layout.addWidget(helpers.create_divider(), 2, 0, 1, 2)
+        server_layout.addWidget(daily_reset_label, 3, 0, 1, 2)
+        server_layout.addWidget(daily_reset_timer_label, 4, 0, 1, 2)
+        server_layout.addWidget(weekly_reset_label, 5, 0, 1, 2)
+        server_layout.addWidget(weekly_reset_timer_label, 6, 0, 1, 2)
 
-        layout.addLayout(server_layout)
-        layout.addItem(QSpacerItem(0, 0, QSizePolicy.Expanding, QSizePolicy.Minimum))
+        daily_reset_label.setAlignment(Qt.AlignCenter)
+        daily_reset_timer_label.setAlignment(Qt.AlignCenter)
+        weekly_reset_label.setAlignment(Qt.AlignCenter)
+        weekly_reset_timer_label.setAlignment(Qt.AlignCenter)
+
+        server_layout.setAlignment(Qt.AlignCenter)
 
         self.threadpool = QtCore.QThreadPool()
         self.threadpool.setMaxThreadCount(2)
@@ -62,8 +133,18 @@ class ServerStatusBar(QWidget):
         self.timer.timeout.connect(self.request_server_status)
         self.timer.start(15000)  # 5000 milliseconds = 5 seconds
         self.request_server_status()
+        self.setLayout(server_layout)
+        self.setMaximumWidth(parent.width * 0.25)
+        self.setSizePolicy(QSizePolicy.Maximum, QSizePolicy.Maximum)
 
-        self.setLayout(layout)
+        self.adjustSize()
+
+    def paintEvent(self, event):
+        # Draw border using QPainter
+        painter = QPainter(self)
+        painter.setRenderHint(QPainter.Antialiasing)  # Enable anti-aliasing
+        painter.setPen(QPen(QColor("#D9D9D9"), 2))
+        painter.drawRoundedRect(self.rect(), 5, 5)
 
     def request_server_status(self):
         if not self.window().isMinimized():
