+++ launcher/version.py
@@ -1,0 +1,1 @@
+version = "v0.2.0"
